+++ src/game/PoolManager.cpp
@@ -417,7 +417,7 @@ void PoolManager::LoadFromDB()
 
     uint32 count = 0;
 
-    barGoLink bar(result->GetRowCount());
+    barGoLink bar((int)result->GetRowCount());
     do
     {
         ++count;
@@ -455,7 +455,7 @@ void PoolManager::LoadFromDB()
     else
     {
 
-        barGoLink bar2(result->GetRowCount());
+        barGoLink bar2((int)result->GetRowCount());
         do
         {
             Field *fields = result->Fetch();
@@ -517,7 +517,7 @@ void PoolManager::LoadFromDB()
     else
     {
 
-        barGoLink bar2(result->GetRowCount());
+        barGoLink bar2((int)result->GetRowCount());
         do
         {
             Field *fields = result->Fetch();
@@ -586,7 +586,7 @@ void PoolManager::LoadFromDB()
     else
     {
 
-        barGoLink bar2( result->GetRowCount() );
+        barGoLink bar2( (int)result->GetRowCount() );
         do
         {
             Field *fields = result->Fetch();
